chore(risk): rename mcpname package to toolref
Route the Codex hook's MCP tool-name attribution through toolref.AttributeTool
instead of a hand-rolled mcp__<server>__<tool> split. No behavior change.
